rename view components
